move handlers into handler namespace
