+++ Peer/Verbs/SubVerbHandler.cs
@@ -1,6 +1,5 @@
 using System;
 using Peer.Apps;
-using Peer.Parsing.CommandLine;
 
 namespace Peer.Verbs;
